% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversions.R
\name{add_visitname_col}
\alias{add_visitname_col}
\title{Adds a mapped visit_name column to a table given the visitplan table}
\usage{
add_visitname_col(table, id = "visit_name", visitplan_table)
}
\arguments{
\item{table}{data.frame with a column mnpvisid to be translated to visit names.}

\item{id}{string This specifies the name of the new id column.}

\item{visitplan_table}{data.frame This is the loaded reference table to make the translation (i.e. visitplan/vp in export archive).}
}
\value{
Returns a data.frame with a new column containing the translated visit names.
}
\description{
Adds a mapped visit_name column to a table given the visitplan table
}
\examples{
\donttest{
bmd_with_visit_name <- add_pat_id_col(table = bmd,
                                      id = "visit_name",
                                      visitplan_table = visitplan)
}

}
