% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframe.R
\name{as.data.frame.secuTrialdata}
\alias{as.data.frame.secuTrialdata}
\title{as.data.frame method for secuTrialdata objects}
\usage{
\method{as.data.frame}{secuTrialdata}(object, meta = TRUE,
  rm_regex = NULL, envir = .GlobalEnv)
}
\arguments{
\item{object}{secuTrialdata object}

\item{meta}{logical should metadata be returned}

\item{rm_regex}{regex syntax to remove from names}

\item{envir}{environment in which to put the data}
}
\value{
each data.frame on the secuTrialdata object is saved to it's own data.frame in the environment
}
\description{
as.data.frame method for secuTrialdata objects
}
\examples{
export_location <- system.file("extdata",
                               "s_export_CSV-xls_CTU05_longnames_sep_ref.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial_export(data_dir = export_location)
# add files to global environment
as.data.frame(sT_export)
# add files to global environment, removing the project name from the file names
as.data.frame(sT_export, rm_regex = "ctu05")
}
