% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assess_form_variable_completeness.R
\name{assess_form_variable_completeness}
\alias{assess_form_variable_completeness}
\title{Asses completeness of data for each variable in a secuTrial export}
\usage{
assess_form_variable_completeness(form, casenodes_table,
  validation_overview, completeness = "allforms", occ_in_vp = 1,
  omit.mnp = TRUE)
}
\arguments{
\item{form}{data.frame Form for which to assess variable completeness
(i.e. a list element returned by load_secuTrial_export).}

\item{casenodes_table}{data.frame The central casenodes record file
(i.e. 'casenodes' from the list returned by load_secuTrial_export).}

\item{validation_overview}{tibble returned by load_validation_overview.}

\item{completeness}{string Specifies if completeness is assessed for all forms ('allforms')
or only for saved forms ('savedforms').}

\item{occ_in_vp}{integer Specifies how often the form occurs in the visit plan
(only relevant if completeness = 'allforms' is set).}

\item{omit.mnp}{boolean Removes variable names from the result table that start with mnp.}
}
\value{
data.frame showing percent completeness per variable.
}
\description{
Variable completeness is defined as percentage of data entered for a variable
when considering all occurrences of the variable throughout the visit plan and
all patients/cases registered in the secuTrial data base. The completeness can
be assessed for 'allforms', which will also take unsaved forms into account, and
for 'savedforms' which will only consider forms that have been actively saved.
In order to asses variable completeness the function requires a loaded secuTrial
validation overview and a loaded secuTrial standard export. Both the validation
overview and the data export should be from the same time (not more than minutes
apart if possible) to ensure data integrity.
Variable completeness is considered based on rules in the secuTrial setup. Thus,
if a variable is not marked as necessary for completeness, this variable will
always be considered as 100 percent complete if only 'savedforms' are assessed.
}
\examples{
\donttest{
assess_form_variable_completeness(form = sT_export$bmd, casenodes_table = sT_export$cn,
                                  validation_overview = val_ovv, completeness = "allforms",
                                  occ_in_vp = 5)
}

}
\seealso{
load_validation_overview, load_secuTrial_export
}
