% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_secuTrial_url.R
\name{build_secuTrial_url}
\alias{build_secuTrial_url}
\title{Compose a secuTrial URL}
\usage{
build_secuTrial_url(server, instance = NA, customer = NA,
  projid = NA, docid = NA)
}
\arguments{
\item{server}{- string containing a server URL}

\item{instance}{- (optional) string containing secuTrial instance name}

\item{customer}{- (optional) string containing secuTrial customer label}

\item{projid}{- (optional) string containing secuTrial project identifier}

\item{docid}{- (optional) secuTrial document/form identifer}
}
\value{
string containing a URL to desired secuTrial page. Currently we provide no guarantee that the returned URL is valid.
}
\description{
Given a Secutrial server URL, and optionally instance, customer, project id, document id and subdoc id,
this function composes a URL to a specific secuTrial instance, customer, form or subform.
}
\details{
If only server URL is provided, the output will point to the secutrial Server
If secuTrial server and instance are provided, the output will point to the secuTrial instance.
If secutrial server, instance and customer id are provided, the output will point to the customer page.
If secuTrial server, instance, customer, project and document id are provided, the output will point to a specific secuTrial form.
}
