% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates_secuTrial.R
\name{dates_secuTrial}
\alias{dates_secuTrial}
\alias{dates_secuTrial.secuTrialdata}
\alias{dates_secuTrial.data.frame}
\alias{dates_secuTrial.character}
\alias{dates_secuTrial.factor}
\alias{dates_secuTrial.integer}
\alias{dates_secuTrial.numeric}
\alias{dates_secuTrial.logical}
\alias{dates_secuTrial.Date}
\title{Methods to handle dates in secuTrial exports}
\usage{
dates_secuTrial(x, ...)

\method{dates_secuTrial}{secuTrialdata}(object)

\method{dates_secuTrial}{data.frame}(data, datevars, format)

\method{dates_secuTrial}{character}(var, format)

\method{dates_secuTrial}{factor}(var, format)

\method{dates_secuTrial}{integer}(var, format)

\method{dates_secuTrial}{numeric}(var, format)

\method{dates_secuTrial}{logical}(var, format)

\method{dates_secuTrial}{Date}(var, format)
}
\arguments{
\item{object}{secuTrialdata object}

\item{data}{data.frame}

\item{datevars}{string consisting of variables with dates}

\item{format}{format of dates (typically taken from \code{object$export_options$date.format})}

\item{var}{date variable to be converted}
}
\value{
same as the original object with date variables converted to Dates.
}
\description{
Methods to handle dates in secuTrial exports
}
\details{
New variables are created appended with \code{.date}. This is a safety mechanism incase NAs are inadvertently introduced.
Methods exist for secuTrialdata objects, data.frames, character, factor, integer and logical classes.
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "s_export_CSV-xls_CTU05_longnames_sep_ref.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- load_secuTrial_export(data_dir = export_location)
# prepare dates
sT_export_dates <- dates_secuTrial(sT_export)
}
