% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary_secuTrial.R
\name{dictionary_secuTrial}
\alias{dictionary_secuTrial}
\title{Data dictionary for secuTrialdata objects}
\usage{
dictionary_secuTrial(x)
}
\arguments{
\item{x}{\code{secuTrialdata} object}
}
\value{
data frame with info on variables
}
\description{
Data dictionary for secuTrialdata objects
}
\examples{
export_location <- system.file("extdata", "sT_exports", "lnames",
                               "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# read all export data
sT_export <- read_secuTrial(data_dir = export_location)
# get dictionary
dictionary_secuTrial(sT_export)
}
