% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_secuTrial.R
\name{diff_secuTrial}
\alias{diff_secuTrial}
\title{Returns differences in the setup of two secuTrialdata objects}
\usage{
diff_secuTrial(x, y)
}
\arguments{
\item{x}{a \code{secuTrialdata} object (the older export)}

\item{y}{a \code{secuTrialdata} object (the newer export)}
}
\description{
During ongoing studies it is possible that changes to the DataCapture interface
             are made. Sometimes these changes may call for adjustments in analysis code.
             It is considered good practice to run \code{diff_secuTrial()} on the last export
             and the current export of a project export to at least make yourself aware of
             potential changes to the setup. If there are changes, the results of this function should
             be interpreted as a first indicator since they may not cover all changed aspects.
             Information is returned on new forms and variables.
             A detailed list of changes can be produced in the FormBuilder with
             "Compare project setup".
}
\details{
If there are differences, \code{diff_secuTrial()} will produce a list of vectors.
         The fist vector informs about new forms and the second vector informs about
         new variables.
}
\examples{
# export location
export_loc <- system.file("extdata", "sT_exports", "longnames",
                          "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                          package = "secuTrialR")
# read export
sT_export <- read_secuTrial(export_loc)

# return scores // in this case on the same export
diff_secuTrial(sT_export, sT_export)

}
