% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/durations_secutrial.R
\name{durations_secuTrial}
\alias{durations_secuTrial}
\title{Methods to handle durations and clock time variable in secuTrial exports}
\usage{
durations_secuTrial(x, ...)
}
\arguments{
\item{object}{secuTrialdata object}

\item{warn}{logical, return warnings}

\item{days_in_month}{numeric, number of days in month for conversion to minutes, defaults to 30.4}

\item{days_in_year}{numeric, number of days in year for conversion to minutes, defaults to 365}
}
\description{
Converts durations to the smallest unit (e.g. minutes for variables recorded as hours:minutes) and formats clock times as HH:MM strings rather than (H)HMM integers, as appropriate.
}
