% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{form_status_counts}
\alias{form_status_counts}
\title{A function to assess the status of forms}
\usage{
form_status_counts(x, ...)
}
\arguments{
\item{object}{secuTrialdata object}
}
\description{
This function returns a data.frame informing on the status of
forms per patient. There is a line for every combination of
form type and patient id. The numbers are occurence counts.
}
\examples{
# prepare path to example export
export_location <- system.file("extdata",
                               "s_export_CSV-xls_CTU05_shortnames_sep_ref.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial(data_dir = export_location)

# get form status
form_status_counts(sT_export)

}
\keyword{completeness}
\keyword{form}
\keyword{status,}
