% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{form_status_summary}
\alias{form_status_summary}
\title{A function to assess the status of forms}
\usage{
form_status_summary(object)
}
\arguments{
\item{object}{secuTrialdata object}
}
\description{
This function returns a data.frame informing on the status of
forms per patient. There is a line for every combination of
form type and patient id. The numbers are occurence counts.
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "s_export_CSV-xls_BMD.zip",
                              package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial_export(data_dir = export_location)

# get form status
form_status_summary(sT_export)

}
\keyword{completeness}
\keyword{form}
\keyword{status,}
