% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_export_options.R
\name{load_export_options}
\alias{load_export_options}
\title{Load secuTrial data export options}
\usage{
load_export_options(data_dir, add_id_name = NULL)
}
\arguments{
\item{data_dir}{string The data_dir specifies the path to the secuTrial data export.}

\item{add_id_name}{string This option needs to be specified if your Add-ID name has been changed in the AdminTool Design setting.}
}
\value{
The function returns a list with the data export options.
}
\description{
The load_export_options reads the secuTrial data export and
determines specific options needed to handle the data properly.
}
\examples{
\donttest{
load_export_options(data_dir = system.file("extdata",
                                           "s_export_CSV-xls_BMD.zip",
                                           package = "secuTrialR"))
}

}
