% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secuTrial.R
\name{load_secuTrial}
\alias{load_secuTrial}
\title{Load secutrial export}
\usage{
load_secuTrial(data_dir, labels = TRUE, factor = TRUE, dates = TRUE)
}
\arguments{
\item{data_dir}{string - location of the export}

\item{labels}{logical - add labels to variables and table}

\item{factor}{logical - convert categorical variables to factor variables (ignored when reference values are not in a seperate table)}

\item{dates}{logical - convert date variables}
}
\value{
secuTrialdata object - a list with one data.frame for each file on the export.
}
\description{
Convenience wrapper for \code{load_secuTrial_export}, \code{label_secuTrial}, \code{factorize_secuTrial} and \code{dates_secuTrial}.
}
\examples{
path <- system.file("extdata", "s_export_CSV-xls_CTU05_longnames_sep_ref.zip", package = "secuTrialR")
d <- load_secuTrial(path)
}
