% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secuTrial_export.R
\name{load_secuTrial_export}
\alias{load_secuTrial_export}
\title{This function loads a secuTrial export}
\usage{
load_secuTrial_export(data_dir, tables = "all", add_id_name = NULL)
}
\arguments{
\item{data_dir}{string The data_dir specifies the path to the secuTrial data export.}

\item{tables}{vector containing the names of the files to be loaded. Default
behavior load all tables. "none" can also be
specified if only patient, centre and export_options
shall be loaded.}

\item{add_id_name}{string This needs to be specified if your Add-ID name has
been changed in the AdminTool Design setting.}
}
\value{
This is a list of data.frames with all the data loaded from the secuTrial export.
        The minimum list will always contain the patient, centre, visitplan and
        export_options data.frames.
}
\description{
This function will always load the casenodes/cn table as patient,
centres/ctr table as centre and visitplan/vp as visitplan.
Further tables to load can be specified. The standard is to load
the entire export. The export options are also loaded and written
into export_options.
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "s_export_CSV-xls_BMD.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- load_secuTrial_export(data_dir = export_location)

}
