% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_validation_overview.R
\name{load_validation_overview}
\alias{load_validation_overview}
\title{This function loads a multi-page secuTrial 'Validation Overview' report into an R tibble.}
\usage{
load_validation_overview(data_dir)
}
\arguments{
\item{data_dir}{Path to the Validation Overview (must be an *.xlsx file).}
}
\value{
tibble with the 'Validation Overview' data.
}
\description{
This function loads a multi-page secuTrial 'Validation Overview' report into an R tibble.
}
\examples{
val_ovv_location <- system.file("extdata", "bmd_validation_overview.xlsx", package = "secuTrialR")
val_ovv <- load_validation_overview(data_dir = val_ovv_location)

}
