% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversions.R
\name{mnppid2mnpaid}
\alias{mnppid2mnpaid}
\title{Find the mnpaid related to an mnppid}
\usage{
mnppid2mnpaid(mnppid, patient_table)
}
\arguments{
\item{mnppid}{The mnppid for which the mnpaid should be retrieved.}

\item{patient_table}{data.frame This is the loaded reference table to make the translation (i.e. casenodes/cn in export archive).}
}
\value{
mnpaid related to the entered mnppid.
}
\description{
Find the mnpaid related to an mnppid
}
\examples{
\donttest{
mnppid2mnpaid(mnppid = 1780, patient_table = patient)
[1] 103
}

}
