% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversions.R
\name{mnpvisid2mnpvislabel}
\alias{mnpvisid2mnpvislabel}
\title{Find the mnpvislabel related to an mnpvisid}
\usage{
mnpvisid2mnpvislabel(mnpvisid, visitplan_table)
}
\arguments{
\item{mnpvisid}{The mnpvisid for which the mnpvislabel should be retrieved.}

\item{visitplan_table}{data.frame This is the loaded reference table to make the translation (i.e. visitplan/vp in export archive).}
}
\value{
mnpvislabel related to the entered mnpvisid.
}
\description{
Find the mnpvislabel related to an mnpvisid
}
\examples{
\donttest{
mnpvisid2mnpvislabel(mnpvisid = 13920, visitplan_table = visitplan)
[1] "#. visit"
}
}
