% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recruitment.R
\name{plot_recruitment}
\alias{plot_recruitment}
\title{Plots the recruitment over time for secuTrialdata objects}
\usage{
plot_recruitment(x, return_data = FALSE)
}
\arguments{
\item{x}{a \code{secuTrialdata} object}

\item{return_data}{logical - return the data used to produce the plot instead of the plot}
}
\description{
secuTrial exports inherently contain the information on which case was
             registered at which point in time. This function makes use of this property
             to plot recruitment over time.
}
\details{
plot_recruitment will return a simple line plot showing recruitment over time
}
\examples{
# export location
expot_loc <- system.file("extdata",
                         "s_export_CSV-xls_CTU05_longnames_sep_ref.zip",
                         package = "secuTrialR")
# read export
sT_export <- read_secuTrial(expot_loc)

# plot recruitment
plot_recruitment(sT_export)

}
