% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.secuTrialdata}
\alias{print.secuTrialdata}
\title{print method for secuTrialdata objects}
\usage{
\method{print}{secuTrialdata}(x)
}
\arguments{
\item{x}{secuTrialdata object as returned by \code{load_secuTrial_export}}
}
\description{
print method for secuTrialdata objects
}
