% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_secuTrial.R
\name{read_secuTrial}
\alias{read_secuTrial}
\title{Load secuTrial export}
\usage{
read_secuTrial(data_dir, labels = TRUE, factor = TRUE, dates = TRUE)
}
\arguments{
\item{data_dir}{string - location of the export}

\item{labels}{logical - add labels to variables and table}

\item{factor}{logical - convert categorical variables to factor variables (ignored when reference values are not in a separate table)}

\item{dates}{logical - convert date variables}
}
\value{
secuTrialdata object - a list with one data.frame for each file on the export.
}
\description{
Convenience wrapper for \code{read_secuTrial_raw}, \code{label_secuTrial}, \code{factorize_secuTrial} and \code{dates_secuTrial}.
}
\examples{
path <- system.file("extdata",
                    "s_export_CSV-xls_CTU05_longnames_sep_ref.zip",
                    package = "secuTrialR")
d <- read_secuTrial(path)
}
