% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversions.R
\name{remove_trailing_bracket}
\alias{remove_trailing_bracket}
\title{Removes a trailing bracket and its contents from a string}
\usage{
remove_trailing_bracket(x)
}
\arguments{
\item{x}{string}
}
\value{
Returns a string without the trailing brackets.
}
\description{
Removes a trailing bracket and its contents from a string
}
\examples{
\donttest{
remove_trailing_bracket("Hospital (Test Study)")
}

}
