% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_secuTrial_export.R
\name{secuTrialdata}
\alias{secuTrialdata}
\alias{load_secuTrial_export}
\alias{print.secuTrialdata}
\alias{plot.secuTrialdata}
\title{This function loads a secuTrial export}
\usage{
load_secuTrial_export(data_dir, tables = "all", add_id_name = NULL)

print(x)

plot(x)
}
\arguments{
\item{data_dir}{string The data_dir specifies the path to the secuTrial data export.}

\item{tables}{vector containing the names of the files to be loaded. Default
behavior load all tables. "none" can also be
specified if only meta tables and export_options
shall be loaded.}

\item{add_id_name}{string This needs to be specified if your Add-ID name has
been changed in the AdminTool Design setting.}

\item{x}{secuTrialdata object as returned by \code{load_secuTrial_export}}
}
\value{
This is a list of class secuTrialdata containing a list of
        export options and data.frames with all the data loaded from
        the secuTrial export.
        The list will contain at least the metadata data.frames and
        export_options list.

data.frame with a row for each table in the export. For each table it
        contains the name, number of rows and columns, an indicator for
        whether the table is a metadata table and the files original name.
}
\description{
This function will always load the full set of meta tables first.
Further tables to load can be specified. The standard is to load
the entire export. The export options are also loaded and written
into export_options.
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "s_export_CSV-xls_BMD.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- load_secuTrial_export(data_dir = export_location)

# Print method
print(sT_export)
# or
sT_export
}
