% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visit_structure.R
\name{visit_structure}
\alias{visit_structure}
\alias{plot.secuTrialvisit}
\title{Get the visit structure of secuTrialdata objects}
\usage{
visit_structure(x)

plot(x)
}
\arguments{
\item{x}{secuTrialdata object}
}
\value{
data.frame with 1 for whether a form (rows) was collected during a particular visit (columns)
}
\description{
Get the visit structure of secuTrialdata objects
}
\note{
Requires a fixed visit structure - an error will be returned for projects without
      a visit structure or one with flexible visits
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "BMD",
                               "s_export_CSV-xls_BMD_short_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial_raw(data_dir = export_location)
\dontrun{
visit_structure(sT_export)
}
\dontrun{
  vs <- visit_structure(sT_export)
  plot(vs)
}
}
