% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_secutrial.R
\name{write_secuTrial}
\alias{write_secuTrial}
\title{Write secuTrial exports to other formats}
\usage{
write_secuTrial(object, ...)
}
\arguments{
\item{object}{secuTrialdata object}

\item{path}{directory where the files should be saved}

\item{format}{format in which to save the export (one of "dta", "sas", "sav", "xpt")}
}
\value{
a list of filenames
}
\description{
Convert the export prepared in R and export it to
             SPSS (sav), Stata (dta) or SAS (sas, xpt)
             using the haven package.
}
\details{
Due to variable naming limitations in other packages, date variables are
         appended with _d (rather than _date), datetime/POSIX variables are appended
         with _dt (rather than _datetime) and factors with _f (rather than _factor).
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "longnames",
                               "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial(data_dir = export_location)
tdir <- tempdir()
write_secuTrial(sT_export, format = "dta", path = tdir)
list.files(tdir)

}
