% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_conversions.R
\name{add_pat_id_col}
\alias{add_pat_id_col}
\title{Adds a mapped pat.id column to a table given the patient table}
\usage{
add_pat_id_col(table, id = "pat.id", patient_table)
}
\arguments{
\item{table}{data.frame with a column mnppid to be translated to pat.id (mnpaid).}

\item{id}{string This specifies the name of the new id column.}

\item{patient_table}{data.frame This is the loaded reference table to make the translation (i.e. casenodes/cn in export archive).}
}
\value{
Returns a data.frame with a new first column containing the translated ids.
}
\description{
Adds a mapped pat.id column to a table given the patient table
}
\examples{
\donttest{
bmd_with_patid <- add_pat_id_col(table = bmd,
                                 id = "pat.id",
                                 patient_table = patient)
}

}
