% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asdataframe.R
\name{as.data.frame.secuTrialdata}
\alias{as.data.frame.secuTrialdata}
\title{as.data.frame method for secuTrialdata objects}
\usage{
\method{as.data.frame}{secuTrialdata}(
  x,
  ...,
  data.frames = NULL,
  meta = FALSE,
  regex = NULL,
  rep = "",
  envir
)
}
\arguments{
\item{x}{\code{secuTrialdata} object}

\item{...}{further parameters}

\item{data.frames}{character vector of data.frame names to turn into data.frames}

\item{meta}{logical should metadata be returned}

\item{regex}{regex syntax to remove from names}

\item{rep}{replacement for regex}

\item{envir}{environment in which to put the data (e.g. \code{.GlobalEnv})}
}
\value{
each data.frame on the \code{secuTrialdata} object is saved to it's own data.frame in the environment
}
\description{
as.data.frame method for secuTrialdata objects
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "sT_exports", "lnames",
                               "s_export_CSV-xls_CTU05_long_ref_miss_en_utf8.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial_raw(data_dir = export_location)
# add files to global environment
as.data.frame(sT_export, envir = .GlobalEnv)
# add files to global environment, removing the project name from the file names
as.data.frame(sT_export, regex = "ctu05", envir = .GlobalEnv)
}
