% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorize.R
\name{factorize}
\alias{factorize}
\alias{factorize_secuTrial}
\alias{factorize_secuTrial.secuTrialdata}
\alias{factorize_secuTrial.data.frame}
\alias{factorize_secuTrial.integer}
\alias{factorize_secuTrial.logical}
\title{Add factors to secuTrialdata objects}
\usage{
factorize_secuTrial(x, ...)

\method{factorize_secuTrial}{secuTrialdata}(object)

\method{factorize_secuTrial}{data.frame}(data, cl, form)

\method{factorize_secuTrial}{integer}(var, lookup)

\method{factorize_secuTrial}{logical}(var, lookup)
}
\arguments{
\item{x}{a \code{secuTrialdata} object}

\item{object}{a \code{secuTrialdata} object}

\item{data}{a \code{data.frame}, usually from within a \code{secuTrialdata} object}

\item{cl}{a \code{data.frame}, usually taken from a \code{secuTrialdata} object, containing at least three variables - 1) \code{column} containing information on which form and variable is to be factorized (formatted as form.variable); 2) \code{code} containing the options that the variable can take as a number; 3) \code{value} contains the text relating to the \code{code} number}

\item{form}{which form you are currently working on (used for filtering \code{cl})}

\item{var}{a variable}

\item{lookup}{a restricted version of cl (filtered based on \code{form}), containing only the rows relevant for \code{var}}
}
\value{
object with extra variables in forms for factors (names are appended with \code{.factor})
}
\description{
Add factors to secuTrialdata objects

Method for secuTrialdata objects

Method for data.frame objects

Method for integer objects

Method for logical objects
}
