% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/completeness.R
\name{form_status_summary}
\alias{form_status_summary}
\title{A function to show summary statistics for form statuses}
\usage{
form_status_summary(x, ...)
}
\arguments{
\item{object}{secuTrialdata object}
}
\description{
This function warps form_status_counts and returns a
data.frame summarizing the statuses for each form.
}
\examples{
# prepare path to example export
export_location <- system.file("extdata",
                               "s_export_CSV-xls_CTU05_shortnames_sep_ref.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- read_secuTrial(data_dir = export_location)

# get form status
form_status_summary(sT_export)

}
\keyword{completeness}
\keyword{form}
\keyword{status,}
