% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/links_secuTrial.R
\name{links_secuTrial}
\alias{links_secuTrial}
\title{Show links between forms}
\usage{
links_secuTrial(object, forms = NULL)
}
\arguments{
\item{object}{secuTrialdata object}

\item{forms}{a regular expression for which forms should be included}
}
\value{
a tikz plot window. We recommend to resize the window and and click view::fit to screen to improve readability. Forms are coloured red, variables are coloured blue.
}
\description{
SecuTrial creates a large number of files and identifiers with which to link files together. Understanding the links can be difficult. This function produces a map linking the forms with common variables.
}
\examples{
# prepare path to example export
export_location <- system.file("extdata", "s_export_CSV-xls_BMD.zip",
                               package = "secuTrialR")
# load all export data
sT_export <- load_secuTrial_export(data_dir = export_location)
# get labels
links_secuTrial(sT_export)
}
