% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_export_table.R
\name{load_export_table}
\alias{load_export_table}
\title{Load an individual file/table from a secuTrial export archive}
\usage{
load_export_table(data_dir, file_name, export_options, add_pat_id = TRUE,
  add_centre = TRUE, patient_table = patient, centre_table = centre)
}
\arguments{
\item{data_dir}{string The data_dir specifies the path to the secuTrial data export.}

\item{file_name}{string The file_name specifies which file to load.}

\item{export_options}{list This can be generated with load_export_options.}

\item{add_pat_id}{boolean If TRUE this will add the pat.id to the table.}

\item{add_centre}{boolean If TRUE this will add the centre name to the table.}

\item{patient_table}{data.frame This data.frame must be created with this function
before any other secuTrial tables are read (exception centre_table).
While creating it this argument is obsolete. (see examples)}

\item{centre_table}{data.frame his data.frame must be created with this function
before any other secuTrial tables are read (exception patient_table).
While creating it this argument is obsolete. (see examples)}
}
\value{
The function returns a data.frame for the data in file_name.
}
\description{
This function can load individual tables from a secuTrial data
export. Before loading any other tables the casenodes/cn
and centres/ctr tables need to be loaded with this function.
Also export_options needs to be generated with load_export_options.
}
\examples{
\donttest{
# load patient (casenodes/cn) table
patient <- load_export_table(data_dir = system.file("extdata",
                                                    "s_export_CSV-xls_BMD.zip",
                                                    package = "secuTrialR"),
                             file_name = "cn.xls",
                             export_options=export_options)

# load centre (centres/ctr) table
centre <- load_export_table(data_dir = system.file("extdata",
                                                   "s_export_CSV-xls_BMD.zip",
                                                   package = "secuTrialR"),
                            file_name = "ctr.xls",
                            export_options=export_options)

# load bone mineral denisty form data
bmd <- load_export_table(data_dir = system.file("extdata",
                                                "s_export_CSV-xls_BMD.zip",
                                                package = "secuTrialR"),
                         file_name = "bmd.xls",
                         export_options = export_options,
                         patient_table = patient,
                         centre_table = centre)
}

}
\seealso{
load_export_options
}
