% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_secuTrial}
\alias{subset_secuTrial}
\title{Subsets a secuTrialdata object}
\usage{
subset_secuTrial(dat, patient = NULL, centre = NULL, exclude = FALSE)
}
\arguments{
\item{dat}{secuTrialdata object containing patient IDs and centre information}

\item{patient}{character vector with a selection of patient IDs (mnpaid) used for subsetting}

\item{centre}{character vector with a selection of centre names (mnpctrname) used for subsetting}

\item{exclude}{boolean which if true excludes given pat_id and centre from dat}
}
\value{
secuTrialdata object containing only the pat_ids and centres that meet the selection criteria.
}
\description{
Given a secuTrial data object, and subsetting parameters,
this function filters the data object to include only the desired patients. Subsetting is possible based on
patients and based on centres. In order to subset based on patients, patient IDs musst be present in the export.
In order to subset based on centres, centre information must be present in the export.
}
\examples{

path <- system.file("extdata", "sT_exports", "subset",
                    "s_export_CSV-xls_CTU05_short_all-info_en.zip",
                    package = "secuTrialR")
sT_export <- read_secuTrial(path)
patients <- c("RPACK-INS-011", "RPACK-INS-014")
centres <- c("Inselspital Bern (RPACK)", "Universitätsspital Basel (RPACK)")

# subset sT_export
sT_export_patients1 <- subset_secuTrial(dat = sT_export, patient = patients)
sT_export_patients2 <- subset_secuTrial(dat = sT_export, patient = patients, exclude = TRUE)
sT_export_centre <- subset_secuTrial(dat = sT_export, centre = centres, exclude = TRUE)

}
